// src/theme/LightTheme.ts
import type { ThemeTypes } from '@/types/themeTypes/ThemeType';

const PurpleTheme: ThemeTypes = {
    name: 'PurpleTheme',
    dark: false,
    variables: {
        'border-color': '#eeeeee',
        'carousel-control-size': 10,
        'border-opacity': 1,
    },
    colors: {
        primary: '#6B30E0',
        secondary: '#49BEFF',
        info: '#539BFF',
        success: '#13DEB9',
        accent: '#FFAB91',
        warning: '#FFAE1F',
        error: '#FA896B',
        muted: '#5a6a85',
        lightprimary: '#ECF2FF',
        lightsecondary: '#E8F7FF',
        lightsuccess: '#E6FFFA',
        lighterror: '#FDEDE8',
        lightwarning: '#FEF5E5',
        textPrimary: '#2A3547',
        textSecondary: '#2A3547',
        borderColor: '#E5EAEF',
        inputBorder: '#000',
        containerBg: '#fff',
        hoverColor: '#f6f9fc',
        surface: '#ffffff',
        'on-surface-variant': '#ffffff',
        grey100: '#F2F6FA',
    }
};
export { PurpleTheme };
